+++ CHANGES.rst
@@ -6,6 +6,7 @@ Improvements:
  * MXCredentials: Create a new data model for it, separated from the CS API response data model (new MXLoginResponse class).
  * Login: Handle well-known data in the login response - MSC1730 (vector-im/riot-ios/issues/2298).
  * Login: Add kMXLoginFlowTypeCAS & kMXLoginFlowTypeSSO.
+ * MXRestClient: Expose acceptableContentTypes.
 
 Bug Fix:
  * Crypto: Fix crash in MXKeyBackup (vector-im/riot-ios/issues/#2281).